requirements.txt
# requirements.txt
pymongo
chromadb
google-generativeai
gradio
python-dotenv
PyMuPDF
jira
